🚚 zb: Move socket out of the `raw` mod
Also remove the now empty `raw` mod.
